Source highligts and muti-taxonomy search
Combine skill vectorestores into one, to enable prediction of skills of diffrent taxonomies simultaniously.

Also check if the tilte of a course is a known esco term, and include all skills that mention this term into the candidate set. This is useful if a course title mentions a esco skill or occupation directly and makes sure all the related skills are taken into account when predictiong relevant skills.

Add a feature to higlight the most probabel source in the input document, that lead to the prediction of a skill. To generate these source highlights add "sources" to "targets" param. Use this with caution, since the current implementation is expensive.